Separate RoomService and Signal.
