+++ apps/room/server/peer.go
@@ -3,16 +3,27 @@ package server
 import (
 	room "github.com/pion/ion/apps/room/proto"
 	"github.com/pion/ion/pkg/util"
-	"github.com/pion/ion/proto/ion"
 )
+
+/*
+string sid = 1;
+string uid = 2;
+string displayName = 3;
+bytes extraInfo = 4;
+Role role = 5;
+string avatar = 6;
+string vendor = 7;
+string token = 8;
+*/
 
 // Peer represents a peer for client
 type Peer struct {
-	uid             string
-	sid             string
-	info            []byte
-	lastStreamEvent *ion.StreamEvent
-	closed          util.AtomicBool
+	uid         string
+	sid         string
+	info        []byte
+	displayName string
+	sig         room.RoomSignal_SignalServer
+	closed      util.AtomicBool
 }
 
 func NewPeer(sid string, uid string, info []byte) *Peer {
@@ -42,16 +53,16 @@ func (p *Peer) SID() string {
 }
 
 func (p *Peer) send(data *room.Reply) error {
-	/*
-		select {
-		case p.sndCh <- data:
-		default:
-			go func() {
-				p.sndCh <- data
-			}()
-		}
-	*/
-	return nil
+	return p.sig.Send(data)
+}
+
+func (p *Peer) sendMediaPresentation(event *room.MediaPresentation) error {
+	data := &room.Reply{
+		Payload: &room.Reply_Presentation{
+			Presentation: event,
+		},
+	}
+	return p.send(data)
 }
 
 func (p *Peer) sendPeerEvent(event *room.ParticipantEvent) error {
