--- apps/room/server/room_manager.go
@@ -1,4 +1,0 @@
-package server
-
-type roomManager struct {
-}
